# DriverService/app/Http/Controllers/DriverController.php
<?php

namespace App\Http\Controllers;

use App\Models\Driver;
use App\Http\Resources\DriverResource;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Http;
use OpenApi\Annotations as OA;

/**
 * @OA\Info(title="Driver Service API", version="1.0")
 */
class DriverController extends Controller
{
    /**
     * @OA\Get(
     *     path="/api/drivers",
     *     summary="List all drivers",
     *     tags={"Drivers"},
     *     @OA\Response(response=200, description="Successful operation")
     * )
     */
    public function index()
    {
        return DriverResource::collection(Driver::all());
    }

    /**
     * @OA\Post(
     *     path="/api/drivers",
     *     summary="Create a new driver",
     *     tags={"Drivers"},
     *     @OA\RequestBody(
     *         @OA\JsonContent(
     *             @OA\Property(property="user_id", type="integer", example=1),
     *             @OA\Property(property="license_number", type="string", example="LIC123456"),
     *             @OA\Property(property="status", type="string", example="available")
     *         )
     *     ),
     *     @OA\Response(response=201, description="Driver created"),
     *     @OA\Response(response=400, description="Invalid user or not a driver")
     * )
     */
    public function store(Request $request)
    {
        $validated = $request->validate([
            'user_id' => 'required|integer|unique:drivers',
            'license_number' => 'required|string|unique:drivers|max:50',
            'status' => 'nullable|string|in:available,on_duty,unavailable',
        ]);

        // Validasi user dari Authentication Service
        $response = Http::get("http://localhost:8000/api/users/{$validated['user_id']}");
        if ($response->failed() || $response['role'] !== 'driver') {
            return response()->json(['error' => 'Invalid user'], 400);
        }

        $driver = Driver::create($validated);
        return new DriverResource($driver);
    }

    /**
     * @OA\Post(
     *     path="/api/drivers/assign",
     *     summary="Assign a driver to a vehicle",
     *     tags={"Drivers"},
     *     @OA\RequestBody(
     *         @OA\JsonContent(
     *             @OA\Property(property="driver_id", type="integer", example=1),
     *             @OA\Property(property="vehicle_id", type="integer", example=1)
     *         )
     *     ),
     *     @OA\Response(response=200, description="Driver assigned to vehicle"),
     *     @OA\Response(response=400, description="Invalid driver or vehicle"),
     *     @OA\Response(response=404, description="Driver or vehicle not found")
     * )
     */
    public function assign(Request $request)
    {
        $validated = $request->validate([
            'driver_id' => 'required|integer',
            'vehicle_id' => 'required|integer'
        ]);

        // Cek driver
        $driver = Driver::find($validated['driver_id']);
        if (!$driver) {
            return response()->json(['error' => 'Driver not found'], 404);
        }
        if ($driver->status !== 'available') {
            return response()->json(['error' => 'Driver is not available'], 400);
        }

        // Validasi vehicle dari Vehicle Service
        $vehicleResponse = Http::get("http://localhost:8002/api/vehicles/{$validated['vehicle_id']}");
        if ($vehicleResponse->failed()) {
            return response()->json(['error' => 'Vehicle not found'], 404);
        }
        if ($vehicleResponse['status'] !== 'available') {
            return response()->json(['error' => 'Vehicle is not available'], 400);
        }

        // Assign vehicle ke driver
        $driver->update([
            'vehicle_id' => $validated['vehicle_id'],
            'status' => 'on_duty'
        ]);
        return new DriverResource($driver);
    }

    /**
     * @OA\Get(
     *     path="/api/drivers/{id}",
     *     summary="Get a driver by ID",
     *     tags={"Drivers"},
     *     @OA\Parameter(name="id", in="path", required=true, @OA\Schema(type="integer")),
     *     @OA\Response(response=200, description="Success operation"),
     *     @OA\Response(response=404, description="Not found")
     * )
     */
    public function show($id)
    {
        $driver = Driver::findOrFail($id);
        return new DriverResource($driver);
    }

    /**
     * @OA\Put(
     *     path="/api/drivers/{id}",
     *     summary="Update a driver",
     *     tags={"Drivers"},
     *     @OA\Parameter(name="id", in="path", required=true, @OA\Schema(type="integer")),
     *     @OA\RequestBody(
     *         @OA\JsonContent(
     *             @OA\Property(property="user_id", type="integer", example=1),
     *             @OA\Property(property="license_number", type="string", example="LIC123456"),
     *             @OA\Property(property="status", type="string", example="available"),
     *             @OA\Property(property="vehicle_id", type="integer", example=1, nullable=true)
     *         )
     *     ),
     *     @OA\Response(response=200, description="Driver updated"),
     *     @OA\Response(response=400, description="Invalid user or vehicle"),
     *     @OA\Response(response=404, description="Driver not found")
     * )
     */
    public function update(Request $request, $id)
    {
        $driver = Driver::findOrFail($id);
        $validated = $request->validate([
            'user_id' => 'required|integer|unique:drivers,user_id,' . $id,
            'license_number' => 'required|string|unique:drivers,license_number,' . $id . '|max:50',
            'status' => 'nullable|string|in:available,on_duty,unavailable',
            'vehicle_id' => 'nullable|integer',
        ]);

        // Validasi user
        $response = Http::get("http://localhost:8000/api/users/{$validated['user_id']}");
        if ($response->failed() || $response['role'] !== 'driver') {
            return response()->json(['error' => 'Invalid user or not a driver'], 400);
        }

        // Validasi vehicle jika ada
        if (!empty($validated['vehicle_id'])) {
            $vehicleResponse = Http::get("http://localhost:8002/api/vehicles/{$validated['vehicle_id']}");
            if ($vehicleResponse->failed()) {
                return response()->json(['error' => 'Vehicle not found'], 400);
            }
            if ($vehicleResponse['status'] !== 'available' && $validated['status'] === 'on_duty') {
                return response()->json(['error' => 'Vehicle is not available'], 400);
            }
        }

        $driver->update($validated);
        return new DriverResource($driver);
    }

    /**
     * @OA\Delete(
     *     path="/api/drivers/{id}",
     *     summary="Delete a driver",
     *     tags={"Drivers"},
     *     @OA\Parameter(name="id", in="path", required=true, @OA\Schema(type="integer")),
     *     @OA\Response(response=204, description="Driver deleted"),
     *     @OA\Response(response=404, description="Driver not found")
     * )
     */
    public function destroy($id)
    {
        $driver = Driver::findOrFail($id);
        $driver->delete();
        return response()->json(null, 204);
    }
}